write-tree: Hash the files
Instead of only printing file name, let's put all files in the object database.
For now we'll print their OID and their name.

Notice that instead of getting one OID to represent a directory we now get a
separate OID for each file, which isn't very useful. Plus, note that the names
of the files aren't stored in the object databse, they are just printed and then
the information is discarded.

So at this stage `write-tree` isn't useful (it just saves a bunch of files as
blobs) but the next change will fix it.

%% METADATA
Id: write-tree-hash-files