merge: Record both parents on commit
Earlier we merged two commits in the working directory, now we will make sure
that a merge commit is created once we commit the merged working directory.

`ugit merge` command will store the branch we want to merge with in a new ref
called *MERGE_HEAD*. So in our previous example:

                    HEAD
                    v
    o---o---o---o---o
         \
          --o---o
                ^
                some-branch

After running `ugit merge some-branch`, *MERGE_HEAD* will be created and point
to *some-branch*. The presence of *MERGE_HEAD* tells ugit that on the next
commit is a merge commit with two parents - HEAD and *MERGE_HEAD*:

                    HEAD
                    v
    o---o---o---o---o
         \
          --o---o < MERGE_HEAD
                ^
                some-branch

So upon `ugit commit`, ugit will create a commit which will look like we want:

    o---o---o---o---o
         \           \
          --o---o-----o < HEAD
                ^
                some-branch

*MERGE_HEAD* can be deleted at this point.

You can see that we changed `base.merge` to set *MERGE_HEAD*, `base.commit` to
take *MERGE_HEAD* into account and `cli.status` to helpfully inform the user
when we're in the middle of a merge.

%% METADATA
Id: merge-record-both-parents