data: Iter over MERGE_HEAD if it exists
Since we've added *MERGE_HEAD* in the last change, it would be useful to visualize
it as well when running `ugit k`. Add it to the list of refs to iterate over.

Since *MERGE_HEAD* may not exist (we might not be in the middle of a merge), add
a check to `iter_refs` to avoid returning non-existing refs.

%% METADATA
Id: iterate-merge-head