merge-base: Compute common ancestor of a commit
This command is called `merge-base`, it will receive two commit OIDs and find
their common ancestor. In other words, it will find the first parent commit that
both of them share. Graphically it looks like this:

        commit C    commit A
        v           v
    o---o---o---o---o
         \
          --o---o
                ^
                commit B

Commit C is the first commmn ancestor of commits A and B.

This command isn't very useful by itself, but we will explain the motivation for
it in the following change. The name `merge-base` might hint that it will we be
related to improving our `merge` command.

Its operation is pretty brute-force: It saves all parents of the first commit to
a list and it iterates over the parents of the second commit in ancestry order
until it reaches a parent that is a parent of the first commit.

%% METADATA
Id: compute-common-ancestor