merge: Fast-forward merge
So far we've always looked at merges as if they combine two separate lines of
development, like so:

                    HEAD
                    v
    o---o---o---o---o
         \
          --o---o
                ^
                some-branch

And the result is:

    o---o---o---o---o
         \           \
          --o---o-----o < HEAD
                ^
                some-branch

But there is an edge case worth talking about - what if HEAD didn't advance in
parallel to the branch, and we have this following situation?

        HEAD
        v
    o---o
         \
          --o---o
                ^
                some-branch

Currently our code doesn't care and when merging HEAD and *some-branch*, it
would create a merge commit like this:

    o---o---------\
         \         \
          --o---o---o < HEAD
                ^
                some-branch

But this commit doesn't carry any information! We could get rid of it and
instead advance HEAD to point the some-branch directly:

    o---o       HEAD
         \      v
          --o---o
                ^
                some-branch

This type of merge is called a "fast-forward" merge.

In order to know if we can perform a fast-forward merge we need to check if HEAD
is an ancestor of the branch that we want to merge into. If the common ancestor
of HEAD and *some-branch* is HEAD itself it means that in particular HEAD is an
ancestor of *some-branch*. That's why in `merge` we're going to check if
`merge_base == HEAD`. If so, we will just update HEAD to the new location and
read that tree.

(There are some cases where you might still want to create a merge commit even
if you can do a fast-forward merge. For example you might want to document the
fact that a merge has occured. We'll ignore this usecase for now.)

%% METADATA
Id: fast-forward-merge