push: Naive implementation
The opposite of `fetch` is `push`. Instead of downloading remote refs and
objects, it uploads objects and syncronize the local refs to the remote refs.
The usecase for this is when you've added some commits and you'd like to update
a remote repository so that it's syncronized with your local version.

Graphically,

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---o

- Remote repository:

              refs/heads/master
              v
      o---o---o

If we would run `ugit push /path/to/remote master` the result would be:

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---o

- Remote repository:

                      refs/heads/master
                      v
      o---o---o---o---o

We'll create a new `push` CLI command, we'll add a `push_object` function to
*data.py*, which would copy a local object to a remote repository and finally we
would tie it together in *remote.py*: Push all objects that are reachable from
the branch to the remote (Find them our beloved `iter_objects_in_commits`) and
synchronize the remote ref to the local ref value.

This implementation is inefficient, because each time it pushes all objects the
the remote repository regardless of objects that exists there. We'll improve it
in following changes.

%% METADATA
Id: push-naive