push: Send only missing objects
As mentioned earlier, copying all objects on each push is inefficient because
the remote might already have some of them. For example, if the situation is:

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---o

- Remote repository:

                  refs/heads/master
                  v
      o---o---o---o

When we call `push` on *master* there's no need to push all 5 commits and
associated objects. It's enough to push the latest commit and the objects that
were modified in that commit.

Instead of pushing all objects in `remote.push()`, let's add a simple check to
determine which objects the remote has:

- Take all remote refs that exist on the remote. Since the remote might have
  refs that point to branches that we didn't pull yet, filter out all refs that
  point to unknown OIDs.
- Collect into a set all objects that are reachable from the known remote OIDs.
- Collect into a set all objects that are reachable from the local branch that
  is being pushed.
- Take the difference between the two sets. This effectively gives us that
  objects that are needed to fully describe the pushed branch but are missing from
  the remote.
- Push those missing objects.

Git has more advanced heuristics to determine what objects should be pushed, but
we're gonna stop here.

%% METADATA
Id: push-only-missing