reset: Move HEAD
`ugit reset` will allow us to move HEAD to an OID of our choice. This is useful
if we want to undo a commit. For example, if we have this situation:

            7fa39 ac120
                v     v
    o-----o-----o-----o
                      ^
                      master <--- HEAD

We can run this command to undo commit *ac120*:

    $ ugit reset 7fa39

This will point master to *7fa49*, effectively erasing the latest (*ac120*)
commit:

            7fa39 ac120
                v     v
    o-----o-----o-----o
                ^
                master <--- HEAD

Note that not only HEAD moved, but also master (Since HEAD is a symbolic ref
that points to master).

Now if we make another commit, we will get this:

            7fa39 ac120
                v     v
    o-----o-----o-----o
                 \
                  ----o
                      ^
                      master <--- HEAD

And commit *ac120* is effectively removed from master.

You might wonder what is the difference between `checkout` and `reset`, since
they both seem to move HEAD to another commit... Couldn't we just use `ugit
checkout 7fa39` to do the same? Well, no. Unlike `reset`, `checkout` moves the
HEAD without dereferencing it (by passing deref=False). So if we would use
checkout we will get:

            7fa39 ac120
                v     v
    o-----o-----o-----o
                ^     ^
              HEAD    master

`checkout` leaves master at the previous place. That's why we need `reset`, to
move the actual branch and not just HEAD.

Note that we can also undo multiple commits if we move HEAD a few commits back,
like this:

      0dda2 7fa39 ac120
          v     v     v
    o-----o-----o-----o
                      ^
                      master <--- HEAD

    $ ugit reset 0dda2

The end result is undoing both 7fa39 and ac120:

      0dda2 7fa39 ac120
          v     v     v
    o-----o-----o-----o
          ^
          master <--- HEAD

Another point about `reset`: Even though I talked about this command like some
kind of undo command, it's important to remember that `reset` just moves HEAD to
a specified state. In practice this can be used for undoing a commit, but
`reset` is not a generic undo command. It can be used to undo any number of
commits from the tip of the branch (by moving HEAD to an earlier commit), but it
can't undo arbitrary commits.

For example if we want to undo *7fa39* only, we can't do it with `reset`:

      0dda2 7fa39 ac120
          v     v     v
    o-----o-----o-----o
                      ^
                      master <--- HEAD

If we move master to *0dda2*, we will undo *ac120* and *7fa39*, which is not the
same as undoing *7fa39* only. Thus `reset` can't be used to remove a commit from
the middle of the history.

%% METADATA
Id: reset-move-head