data: Allow git directory change
We're going to implement repository syncronization soon (`fetch`, `push`). In
preparation for that, we need to allow ugit to change the current `GIT_DIR`, so
that we can temporarily look at different repositories while syncronizing.

Let's use `contextmanager` to allow the change to be performed in a `with`
statement so that directory changes are easily stackable and revertible.

%% METADATA
Id: change-git-directory