diff: Add --cached option and take index into account
Let's add another mode to `ugit diff`. Previosly the command would diff the
working tree and another commit. Now that we have the index, let's introduce a
few more useful modes to `diff`:

- If no arguments are provided, diff from the index to the working directory.
This way you can quickly see unstaged changes.

- If `--cached` was provided, diff from HEAD to the index. This way you can
quickly see which changes are going to be commited.

- If a specific commit was provided, diff from the commit to the index or
working directory (depending on whether --cached was provided).

Unfortunately `--cached` is an obscure name for a flag, but that's what real Git
uses so we're going to stick to conventions.

%% METADATA
Id: diff-use-index