merge: Create command
Up until now we have implemented good support for developing different branches
of code in parallel. We could easily create new branches and switch between
them. Now let's bring the parallel branches back together. This is known as
merging.

This is useful if multiple people develop code in parallel and they want to
unite their changed into one commit. For example if we have:

                    HEAD
                    v
    o---o---o---o---o
         \
          --o---o
                ^
                some-branch

And we merge HEAD into *some-branch*, we should get:

    o---o---o---o---o
         \           \
          --o---o-----o < HEAD
                ^
                some-branch

This is the big picture and the end goal, it will take us some time to get there
in small steps. Please keep that picture in mind as we implement `merge`.

In this change we've just created an empty `merge` command to get us started.

%% METADATA
Id: merge-command